+++ CHANGELOG.rst
@@ -6,6 +6,18 @@ CHANGELOG
 0.8.2
 -----
 
+* ``SolrQuerySet`` now supports Solr grouping via new `group`
+  method and `GroupedResponse`
+* New class method `prep_index_chunk` on ``Indexable`` class, to support
+  prefetching related objects when iterating over Django querysets for indexing
+* Include django view mixins in sphinx documentation  
+* Dropped support for python 3.6; added python 3.9
+* Dropped support for Django 2.2; added Django 3.2
+* No longer tested against Solr 6.6
+
+0.8.2
+-----
+
 * When subclassing ``SolrQuerySet``, result documents can now be customized by extending ``get_result_document``
 
 0.8.1